[Fix] Fix k8s application type can not render (#3202)
