+++ dinky-web/src/pages/DevOps/JobDetail/JobOperator/components/EditJobInstanceForm.tsx
@@ -17,8 +17,7 @@
  *
  */
 
-import { ClusterType } from '@/pages/RegCenter/Cluster/constants';
-import { CLUSTER_INSTANCE_TYPE } from '@/pages/RegCenter/Cluster/Instance/components/contants';
+import { CLUSTER_INSTANCE_TYPE } from '@/pages/RegCenter/Cluster/constants';
 import { validatorJMHAAdderess } from '@/pages/RegCenter/Cluster/Instance/components/function';
 import { handleAddOrUpdate } from '@/services/BusinessCrud';
 import { API_CONSTANTS } from '@/services/endpoints';
@@ -85,7 +84,7 @@ const EditJobInstanceForm = (props: {
         name='type'
         label={l('rc.ci.type')}
         disabled
-        options={CLUSTER_INSTANCE_TYPE([ClusterType.YARN_APPLICATION])}
+        options={CLUSTER_INSTANCE_TYPE()}
         rules={[{ required: true, message: l('rc.ci.typePlaceholder') }]}
         placeholder={l('rc.ci.typePlaceholder')}
       />
