+++ dinky-web/src/pages/RegCenter/Cluster/Configuration/components/contants.tsx
@@ -18,30 +18,7 @@
  */
 
 import { FormConfig } from '@/pages/RegCenter/Cluster/Configuration/components/data';
-import { ClusterType } from '@/pages/RegCenter/Cluster/constants';
 import { l } from '@/utils/intl';
-import { DefaultOptionType } from 'rc-select/es/Select';
-
-/**
- * Cluster config type
- */
-export const CLUSTER_CONFIG_TYPE: DefaultOptionType[] = [
-  {
-    value: ClusterType.YARN,
-    label: 'Flink On Yarn',
-    key: ClusterType.YARN
-  },
-  {
-    value: ClusterType.KUBERNETES_NATIVE,
-    label: 'Kubernetes Native',
-    key: ClusterType.KUBERNETES_NATIVE
-  },
-  {
-    value: ClusterType.KUBERNETES_OPERATOR,
-    label: 'Kubernetes Operator',
-    key: ClusterType.KUBERNETES_OPERATOR
-  }
-];
 
 export const FLINK_CONFIG_LIST: FormConfig[] = [
   {
@@ -79,52 +56,6 @@ export const FLINK_CONFIG_LIST: FormConfig[] = [
     label: l('rc.cc.ckpDir'),
     placeholder: l('rc.cc.ckpDir'),
     tooltip: l('rc.cc.ckpDirHelp')
-  }
-];
-
-export const DOCKER_CONFIG_LIST: FormConfig[] = [
-  {
-    name: 'dinky.remote.addr',
-    label: l('rc.cc.docker.dinky.addr'),
-    placeholder: l('rc.cc.docker.dinky.addrHelp'),
-    tooltip: l('rc.cc.docker.dinky.addrHelp'),
-    defaultValue: '127.0.0.1:8888'
-  },
-  {
-    name: 'docker.instance',
-    label: l('rc.cc.docker.instance'),
-    placeholder: l('rc.cc.docker.instanceHelp'),
-    tooltip: l('rc.cc.docker.instanceHelp')
-  },
-  {
-    name: 'docker.registry.url',
-    label: l('rc.cc.docker.url'),
-    placeholder: l('rc.cc.docker.urlHelp'),
-    tooltip: l('rc.cc.docker.urlHelp')
-  },
-  {
-    name: 'docker.registry.username',
-    label: l('rc.cc.docker.username'),
-    placeholder: l('rc.cc.docker.usernameHelp'),
-    tooltip: l('rc.cc.docker.usernameHelp')
-  },
-  {
-    name: 'docker.registry.password',
-    label: l('rc.cc.docker.password'),
-    placeholder: l('rc.cc.docker.passwordHelp'),
-    tooltip: l('rc.cc.docker.passwordHelp')
-  },
-  {
-    name: 'docker.image.tag',
-    label: l('rc.cc.docker.tag'),
-    placeholder: l('rc.cc.docker.tagHelp'),
-    tooltip: l('rc.cc.docker.tagHelp')
-  },
-  {
-    name: 'docker.image.dockerfile',
-    label: l('rc.cc.docker.file'),
-    placeholder: l('rc.cc.docker.fileHelp'),
-    tooltip: l('rc.cc.docker.fileHelp')
   }
 ];
