+++ dinky-web/src/pages/RegCenter/Cluster/Instance/components/InstanceModal/InstanceForm/index.tsx
@@ -17,8 +17,7 @@
  *
  */
 
-import { ClusterType } from '@/pages/RegCenter/Cluster/constants';
-import { CLUSTER_INSTANCE_TYPE } from '@/pages/RegCenter/Cluster/Instance/components/contants';
+import { ClusterType, CLUSTER_INSTANCE_TYPE } from '@/pages/RegCenter/Cluster/constants';
 import { validatorJMHAAdderess } from '@/pages/RegCenter/Cluster/Instance/components/function';
 import { Cluster } from '@/types/RegCenter/data.d';
 import { l } from '@/utils/intl';
@@ -65,7 +64,11 @@ const InstanceForm: React.FC<InstanceFormProps> = (props) => {
             label={l('rc.ci.type')}
             width='sm'
             disabled={values && values.autoRegisters}
-            options={CLUSTER_INSTANCE_TYPE([ClusterType.YARN_APPLICATION])}
+            options={CLUSTER_INSTANCE_TYPE([
+              ClusterType.YARN_APPLICATION,
+              ClusterType.KUBERNETES_OPERATOR,
+              ClusterType.KUBERNETES_APPLICATION
+            ])}
             rules={[{ required: true, message: l('rc.ci.typePlaceholder') }]}
             placeholder={l('rc.ci.typePlaceholder')}
           />
